refactor(server): #125 — extract publish/subscribe handlers into routes/messaging.rs
